ingest_api_data/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-pubsub
requests
aiohttp